修正esp_proposals = sm4cbc-sm3-sm2dh配置时，负载明文问题
https://github.com/zhangke5959/strongswan/issues/11

+++ src/libstrongswan/plugins/gmalg/gmalg/private_include/sm4.h
@@ -9,9 +9,9 @@ typedef struct {
 	u32 iv[16];
 } sm4_ctx;
 
-void sm4_ecb_encrypt(sm4_ctx *ctx, u8 *key, u8 *in, u8 len, u8 *out);
-void sm4_ecb_decrypt(sm4_ctx *ctx, u8 *key, u8 *in, u8 len, u8 *out);
-void sm4_cbc_encrypt(sm4_ctx *ctx, u8 *key, u8 *iv, u8 *in, u8 len, u8 *out);
-void sm4_cbc_decrypt(sm4_ctx *ctx, u8 *key, u8 *iv, u8 *in, u8 len, u8 *out);
+void sm4_ecb_encrypt(sm4_ctx *ctx, u8 *key, u8 *in, u32 len, u8 *out);
+void sm4_ecb_decrypt(sm4_ctx *ctx, u8 *key, u8 *in, u32 len, u8 *out);
+void sm4_cbc_encrypt(sm4_ctx *ctx, u8 *key, u8 *iv, u8 *in, u32 len, u8 *out);
+void sm4_cbc_decrypt(sm4_ctx *ctx, u8 *key, u8 *iv, u8 *in, u32 len, u8 *out);
 
 #endif /* _SM4_H_ */
